bugfix: parse ISO date with local timezone in js
For some weird reason browsers interpret ISO strings with missing timezone as UTC if only date portion is present, but as local time if both date and time are present.

In this case, we need to load the date as local time, so we need to append "T00:00:00" to it.